pyproject.toml
[project]
name = "batching-security-checker"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "onnx (>=1.17.0,<2.0.0)",
    "absl-py (>=2.2.0,<3.0.0)",
    "numpy (>=2.2.3,<3.0.0)",
    "attrs (>=25.3.0,<26.0.0)",
    "cattrs (>=24.1.3,<25.0.0)",
    "jax (>=0.6.0,<0.7.0)",
    "rich (>=14.0.0,<15.0.0)",
    "sentencepiece (>=0.2.0,<0.3.0)",
    "jaxtyping (>=0.3.2,<0.4.0)",
    "onnxruntime (>=1.21.1,<2.0.0)",
    "jaxonnxruntime @ git+https://github.com/google/jaxonnxruntime.git",
    "huggingface-hub[cli] (>=0.30.2,<0.31.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "ipykernel (>=6.29.5,<7.0.0)",
    "tinydb (>=4.8.2,<5.0.0)",
    "plotly (>=6.0.1,<7.0.0)",
    "nbformat (>=5.10.4,<6.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
